// planning-idsdp-rms-frontend/src/app/main/modules/rpm/models/AgreementJamanatnamaFormModel.ts
export class AgreementJamanatnamaFormModel {
    agreementId: string;
    guarantorName: string;
    fathersName: string;
    presentAddress: string;
    permanentAddress: string;
    emailAddress: string;
    nidNumber: string;
    mobileNumber: string;
    firstPage: string;
    agreementWithResearcherId: any;
    secondPage: string;
    thirdPage: string;
    fourthPage: string;
    isEditable: boolean = true;
    tinNo: string;
    motherName: string;
    refundPeriod: number;
}